+++ e2e/sveltekit/package.json
@@ -8,9 +8,10 @@
     "api": "cross-env TZ=utc e2e-api",
     "dev": "concurrently \"pnpm run web\" \"pnpm run api\" -n \"web,api\" -c \"green,magenta\"",
     "build": "vite build",
-    "build:": "cd ../../ && run build && cd -",
+    "build:": "cd ../../ && ((run build && cd -) || (cd - && exit 1))",
     "build:dev": "pnpm build: && pnpm dev",
     "build:test": "pnpm build: && pnpm test",
+    "build:build": "pnpm build: && pnpm test",
     "package": "svelte-kit package",
     "previewWeb": "vite preview --port 3007",
     "preview": "concurrently \"pnpm run previewWeb\" \"pnpm run api\" -n \"web,api\" -c \"green,magenta\"",
