+++ site/src/routes/_diagrams/rendered/extract.svelte
@@ -1,1 +1,1 @@
-<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 691.2265625 144" style="max-width: 691.2265625px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674937675548" {...$$props} ><style>#mermaid-1674937675548{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674937675548 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674937675548 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674937675548 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674937675548 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674937675548 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674937675548 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674937675548 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674937675548 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674937675548 .marker.cross{stroke:var(--contrast);}#mermaid-1674937675548 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674937675548 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674937675548 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675548 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675548 .label text,#mermaid-1674937675548 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674937675548 .node rect,#mermaid-1674937675548 .node circle,#mermaid-1674937675548 .node ellipse,#mermaid-1674937675548 .node polygon,#mermaid-1674937675548 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674937675548 .node .label{text-align:center;}#mermaid-1674937675548 .node.clickable{cursor:pointer;}#mermaid-1674937675548 .arrowheadPath{fill:undefined;}#mermaid-1674937675548 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674937675548 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674937675548 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674937675548 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674937675548 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674937675548 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675548 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675548 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674937675548 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674937675548 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-extensions LE-include" id="L-extensions-include-0" d="M90.398,19.5L94.565,19.5C98.732,19.5,107.065,19.5,115.398,19.5C123.732,19.5,132.065,19.5,136.232,19.5L140.398,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-include LE-exclude" id="L-include-exclude-0" d="M207.727,19.5L211.893,19.5C216.06,19.5,224.393,19.5,232.727,19.5C241.06,19.5,249.393,19.5,253.56,19.5L257.727,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-exclude LE-extract" id="L-exclude-extract-0" d="M328.492,19.5L332.659,19.5C336.826,19.5,345.159,19.5,353.492,19.5C361.826,19.5,370.159,19.5,374.326,19.5L378.492,19.5"></path><path style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-extract LE-collect" id="L-extract-collect-0" d="M531.594,19.5L535.76,19.5C539.927,19.5,548.26,19.5,557.764,22.807C567.267,26.114,577.941,32.727,583.278,36.034L588.614,39.341"></path><path style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-schema LE-collect" id="L-schema-collect-0" d="M489.32,108.5L500.533,108.5C511.745,108.5,534.169,108.5,550.718,105.193C567.267,101.886,577.941,95.273,583.278,91.966L588.614,88.659"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(45.19921875, 19.5)" id="flowchart-extensions-37" class="node default default"><rect height="39" width="90.3984375" y="-19.5" x="-45.19921875" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-37.69921875, -12)" style="" class="label"><foreignObject height="24" width="75.3984375"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">extensions</span></div></foreignObject></g></g><g transform="translate(174.0625, 19.5)" id="flowchart-include-38" class="node default default"><rect height="39" width="67.328125" y="-19.5" x="-33.6640625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-26.1640625, -12)" style="" class="label"><foreignObject height="24" width="52.328125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">include</span></div></foreignObject></g></g><g transform="translate(293.109375, 19.5)" id="flowchart-exclude-40" class="node default default"><rect height="39" width="70.765625" y="-19.5" x="-35.3828125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-27.8828125, -12)" style="" class="label"><foreignObject height="24" width="55.765625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">exclude</span></div></foreignObject></g></g><g transform="translate(455.04296875, 19.5)" id="flowchart-extract-42" class="node default default"><rect height="39" width="153.1015625" y="-19.5" x="-76.55078125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-69.05078125, -12)" style="" class="label"><foreignObject height="24" width="138.1015625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">extract_documents</span></div></foreignObject></g></g><g transform="translate(628.41015625, 64)" id="flowchart-collect-44" class="node default default"><circle height="63" width="93.6328125" r="46.81640625" ry="0" rx="0" style=""></circle><g transform="translate(-39.31640625, -24)" style="" class="label"><foreignObject height="48" width="78.6328125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">Parse <br> Documents</span></div></foreignObject></g></g><g transform="translate(455.04296875, 108.5)" id="flowchart-schema-45" class="node default default"><rect height="39" width="68.5546875" y="-19.5" x="-34.27734375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-26.77734375, -12)" style="" class="label"><foreignObject height="24" width="53.5546875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">schema</span></div></foreignObject></g></g></g></g></g></svg>
+<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 689.46875 144" style="max-width: 689.46875px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674954534871" {...$$props} ><style>#mermaid-1674954534871{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674954534871 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674954534871 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674954534871 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674954534871 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674954534871 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674954534871 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674954534871 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674954534871 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674954534871 .marker.cross{stroke:var(--contrast);}#mermaid-1674954534871 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674954534871 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674954534871 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534871 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534871 .label text,#mermaid-1674954534871 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674954534871 .node rect,#mermaid-1674954534871 .node circle,#mermaid-1674954534871 .node ellipse,#mermaid-1674954534871 .node polygon,#mermaid-1674954534871 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674954534871 .node .label{text-align:center;}#mermaid-1674954534871 .node.clickable{cursor:pointer;}#mermaid-1674954534871 .arrowheadPath{fill:undefined;}#mermaid-1674954534871 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674954534871 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674954534871 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674954534871 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674954534871 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674954534871 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534871 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534871 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674954534871 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674954534871 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-extensions LE-include" id="L-extensions-include-0" d="M91.5,19.5L95.667,19.5C99.833,19.5,108.167,19.5,116.5,19.5C124.833,19.5,133.167,19.5,137.333,19.5L141.5,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-include LE-exclude" id="L-include-exclude-0" d="M207.203,19.5L211.37,19.5C215.536,19.5,223.87,19.5,232.203,19.5C240.536,19.5,248.87,19.5,253.036,19.5L257.203,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-exclude LE-extract" id="L-exclude-extract-0" d="M327.359,19.5L331.526,19.5C335.693,19.5,344.026,19.5,352.359,19.5C360.693,19.5,369.026,19.5,373.193,19.5L377.359,19.5"></path><path style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-extract LE-collect" id="L-extract-collect-0" d="M527.547,19.5L531.714,19.5C535.88,19.5,544.214,19.5,553.716,22.754C563.219,26.009,573.89,32.518,579.226,35.772L584.562,39.026"></path><path style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-schema LE-collect" id="L-schema-collect-0" d="M487.969,108.5L498.732,108.5C509.495,108.5,531.021,108.5,547.12,105.246C563.219,101.991,573.89,95.482,579.226,92.228L584.562,88.974"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(45.75, 19.5)" id="flowchart-extensions-37" class="node default default"><rect height="39" width="91.5" y="-19.5" x="-45.75" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-38.25, -12)" style="" class="label"><foreignObject height="24" width="76.5"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">extensions</span></div></foreignObject></g></g><g transform="translate(174.3515625, 19.5)" id="flowchart-include-38" class="node default default"><rect height="39" width="65.703125" y="-19.5" x="-32.8515625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-25.3515625, -12)" style="" class="label"><foreignObject height="24" width="50.703125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">include</span></div></foreignObject></g></g><g transform="translate(292.28125, 19.5)" id="flowchart-exclude-40" class="node default default"><rect height="39" width="70.15625" y="-19.5" x="-35.078125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-27.578125, -12)" style="" class="label"><foreignObject height="24" width="55.15625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">exclude</span></div></foreignObject></g></g><g transform="translate(452.453125, 19.5)" id="flowchart-extract-42" class="node default default"><rect height="39" width="150.1875" y="-19.5" x="-75.09375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-67.59375, -12)" style="" class="label"><foreignObject height="24" width="135.1875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">extract_documents</span></div></foreignObject></g></g><g transform="translate(625.5078125, 64)" id="flowchart-collect-44" class="node default default"><circle height="63" width="95.921875" r="47.9609375" ry="0" rx="0" style=""></circle><g transform="translate(-40.4609375, -24)" style="" class="label"><foreignObject height="48" width="80.921875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">Parse <br> Documents</span></div></foreignObject></g></g><g transform="translate(452.453125, 108.5)" id="flowchart-schema-45" class="node default default"><rect height="39" width="71.03125" y="-19.5" x="-35.515625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-28.015625, -12)" style="" class="label"><foreignObject height="24" width="56.03125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">schema</span></div></foreignObject></g></g></g></g></g></svg>
