+++ site/src/routes/_diagrams/rendered/generate.svelte
@@ -1,1 +1,1 @@
-<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 732.2578125 500" style="max-width: 732.2578125px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674937675578" {...$$props} ><style>#mermaid-1674937675578{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674937675578 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674937675578 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674937675578 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674937675578 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674937675578 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674937675578 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674937675578 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674937675578 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674937675578 .marker.cross{stroke:var(--contrast);}#mermaid-1674937675578 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674937675578 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674937675578 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675578 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675578 .label text,#mermaid-1674937675578 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674937675578 .node rect,#mermaid-1674937675578 .node circle,#mermaid-1674937675578 .node ellipse,#mermaid-1674937675578 .node polygon,#mermaid-1674937675578 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674937675578 .node .label{text-align:center;}#mermaid-1674937675578 .node.clickable{cursor:pointer;}#mermaid-1674937675578 .arrowheadPath{fill:undefined;}#mermaid-1674937675578 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674937675578 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674937675578 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674937675578 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674937675578 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674937675578 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675578 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675578 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674937675578 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674937675578 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-t_before_g LE-gen_sq_center" id="L-t_before_g-gen_sq_center-0" d="M363.457,39L363.457,43.167C363.457,47.333,363.457,55.667,363.457,64C363.457,72.333,363.457,80.667,363.457,84.833L363.457,89"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_center LE-gen_sq_left" id="L-gen_sq_center-gen_sq_left-0" d="M355.957,97.918L317.978,105.098C279.998,112.279,204.039,126.639,166.06,141.986C128.08,157.333,128.08,173.667,128.08,181.833L128.08,190"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_left LE-index_file" id="L-gen_sq_left-index_file-0" d="M120.58,201.429L107.673,208.191C94.766,214.953,68.951,228.476,56.044,239.405C43.137,250.333,43.137,258.667,43.137,262.833L43.137,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_left LE-graphql_tag_return" id="L-gen_sq_left-graphql_tag_return-0" d="M135.58,201.429L148.487,208.191C161.395,214.953,187.209,228.476,200.116,239.405C213.023,250.333,213.023,258.667,213.023,262.833L213.023,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_center LE-artifact_data" id="L-gen_sq_center-artifact_data-0" d="M363.457,104L363.457,110.167C363.457,116.333,363.457,128.667,363.457,141C363.457,153.333,363.457,165.667,363.457,171.833L363.457,178"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-artifact_data LE-artifact_hash" id="L-artifact_data-artifact_hash-0" d="M363.457,217L363.457,221.167C363.457,225.333,363.457,233.667,363.457,242C363.457,250.333,363.457,258.667,363.457,262.833L363.457,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-artifact_hash LE-artifact_selection" id="L-artifact_hash-artifact_selection-0" d="M363.457,306L363.457,310.167C363.457,314.333,363.457,322.667,363.457,331C363.457,339.333,363.457,347.667,363.457,351.833L363.457,356"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-artifact_selection LE-artifact_e" id="L-artifact_selection-artifact_e-0" d="M363.457,395L363.457,399.167C363.457,403.333,363.457,411.667,363.457,420C363.457,428.333,363.457,436.667,363.457,440.833L363.457,445"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_center LE-gen_sq_right" id="L-gen_sq_center-gen_sq_right-0" d="M370.957,97.944L408.217,105.12C445.477,112.296,519.997,126.648,557.257,141.991C594.518,157.333,594.518,173.667,594.518,181.833L594.518,190"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_right LE-transform_runtime" id="L-gen_sq_right-transform_runtime-0" d="M587.018,201.556L574.554,208.297C562.091,215.038,537.165,228.519,524.701,239.426C512.238,250.333,512.238,258.667,512.238,262.833L512.238,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_right LE-generate" id="L-gen_sq_right-generate-0" d="M602.018,201.556L614.481,208.297C626.944,215.038,651.87,228.519,664.334,239.426C676.797,250.333,676.797,258.667,676.797,262.833L676.797,267"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g transform="translate(363.45703125, 141)" class="edgeLabel"><g transform="translate(-71.75, -12)" class="label"><foreignObject height="24" width="143.5"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel">For every document</span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(363.45703125, 19.5)" id="flowchart-t_before_g-119" class="node default default"><rect height="39" width="212.9765625" y="-19.5" x="-106.48828125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-98.98828125, -12)" style="" class="label"><foreignObject height="24" width="197.9765625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_before_generate</span></div></foreignObject></g></g><g transform="translate(363.45703125, 96.5)" id="flowchart-gen_sq_center-120" class="node default default"><rect height="15" width="15" y="-7.5" x="-7.5" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(0, 0)" style="" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel"></span></div></foreignObject></g></g><g transform="translate(128.080078125, 197.5)" id="flowchart-gen_sq_left-122" class="node default default"><rect height="15" width="15" y="-7.5" x="-7.5" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(0, 0)" style="" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel"></span></div></foreignObject></g></g><g transform="translate(43.13671875, 286.5)" id="flowchart-index_file-125" class="node default default"><rect height="39" width="86.2734375" y="-19.5" x="-43.13671875" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-35.63671875, -12)" style="" class="label"><foreignObject height="24" width="71.2734375"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">index_file</span></div></foreignObject></g></g><g transform="translate(213.0234375, 286.5)" id="flowchart-graphql_tag_return-128" class="node default default"><rect height="39" width="153.5" y="-19.5" x="-76.75" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-69.25, -12)" style="" class="label"><foreignObject height="24" width="138.5"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">graphql_tag_return</span></div></foreignObject></g></g><g transform="translate(363.45703125, 197.5)" id="flowchart-artifact_data-131" class="node default default"><rect height="39" width="109.578125" y="-19.5" x="-54.7890625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-47.2890625, -12)" style="" class="label"><foreignObject height="24" width="94.578125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">artifact_data</span></div></foreignObject></g></g><g transform="translate(363.45703125, 286.5)" id="flowchart-artifact_hash-133" class="node default default"><rect height="39" width="47.3671875" y="-19.5" x="-23.68359375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-16.18359375, -12)" style="" class="label"><foreignObject height="24" width="32.3671875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">hash</span></div></foreignObject></g></g><g transform="translate(363.45703125, 375.5)" id="flowchart-artifact_selection-135" class="node default default"><rect height="39" width="142.3125" y="-19.5" x="-71.15625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-63.65625, -12)" style="" class="label"><foreignObject height="24" width="127.3125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">artifact_selection</span></div></foreignObject></g></g><g transform="translate(363.45703125, 464.5)" id="flowchart-artifact_e-137" class="node default default"><rect height="39" width="103.890625" y="-19.5" x="-51.9453125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-44.4453125, -12)" style="" class="label"><foreignObject height="24" width="88.890625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">artifact_end</span></div></foreignObject></g></g><g transform="translate(594.517578125, 197.5)" id="flowchart-gen_sq_right-140" class="node default default"><rect height="15" width="15" y="-7.5" x="-7.5" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(0, 0)" style="" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel"></span></div></foreignObject></g></g><g transform="translate(512.23828125, 286.5)" id="flowchart-transform_runtime-143" class="node default default"><rect height="39" width="150.1953125" y="-19.5" x="-75.09765625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-67.59765625, -12)" style="" class="label"><foreignObject height="24" width="135.1953125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_runtime</span></div></foreignObject></g></g><g transform="translate(676.796875, 286.5)" id="flowchart-generate-145" class="node default default"><rect height="39" width="78.921875" y="-19.5" x="-39.4609375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-31.9609375, -12)" style="" class="label"><foreignObject height="24" width="63.921875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">generate</span></div></foreignObject></g></g></g></g></g></svg>
+<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 713.09375 500" style="max-width: 713.09375px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674954534922" {...$$props} ><style>#mermaid-1674954534922{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674954534922 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674954534922 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674954534922 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674954534922 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674954534922 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674954534922 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674954534922 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674954534922 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674954534922 .marker.cross{stroke:var(--contrast);}#mermaid-1674954534922 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674954534922 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674954534922 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534922 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534922 .label text,#mermaid-1674954534922 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674954534922 .node rect,#mermaid-1674954534922 .node circle,#mermaid-1674954534922 .node ellipse,#mermaid-1674954534922 .node polygon,#mermaid-1674954534922 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674954534922 .node .label{text-align:center;}#mermaid-1674954534922 .node.clickable{cursor:pointer;}#mermaid-1674954534922 .arrowheadPath{fill:undefined;}#mermaid-1674954534922 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674954534922 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674954534922 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674954534922 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674954534922 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674954534922 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534922 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534922 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674954534922 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674954534922 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-t_before_g LE-gen_sq_center" id="L-t_before_g-gen_sq_center-0" d="M357.68,39L357.68,43.167C357.68,47.333,357.68,55.667,357.68,64C357.68,72.333,357.68,80.667,357.68,84.833L357.68,89"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_center LE-gen_sq_left" id="L-gen_sq_center-gen_sq_left-0" d="M350.18,97.931L312.568,105.109C274.957,112.288,199.734,126.644,162.123,141.989C124.512,157.333,124.512,173.667,124.512,181.833L124.512,190"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_left LE-index_file" id="L-gen_sq_left-index_file-0" d="M117.012,201.511L104.394,208.259C91.776,215.007,66.54,228.504,53.923,239.419C41.305,250.333,41.305,258.667,41.305,262.833L41.305,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_left LE-graphql_tag_return" id="L-gen_sq_left-graphql_tag_return-0" d="M132.012,201.511L144.63,208.259C157.247,215.007,182.483,228.504,195.101,239.419C207.719,250.333,207.719,258.667,207.719,262.833L207.719,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_center LE-artifact_data" id="L-gen_sq_center-artifact_data-0" d="M357.68,104L357.68,110.167C357.68,116.333,357.68,128.667,357.68,141C357.68,153.333,357.68,165.667,357.68,171.833L357.68,178"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-artifact_data LE-artifact_hash" id="L-artifact_data-artifact_hash-0" d="M357.68,217L357.68,221.167C357.68,225.333,357.68,233.667,357.68,242C357.68,250.333,357.68,258.667,357.68,262.833L357.68,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-artifact_hash LE-artifact_selection" id="L-artifact_hash-artifact_selection-0" d="M357.68,306L357.68,310.167C357.68,314.333,357.68,322.667,357.68,331C357.68,339.333,357.68,347.667,357.68,351.833L357.68,356"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-artifact_selection LE-artifact_e" id="L-artifact_selection-artifact_e-0" d="M357.68,395L357.68,399.167C357.68,403.333,357.68,411.667,357.68,420C357.68,428.333,357.68,436.667,357.68,440.833L357.68,445"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_center LE-gen_sq_right" id="L-gen_sq_center-gen_sq_right-0" d="M365.18,98.252L395.674,105.377C426.169,112.502,487.159,126.751,517.654,142.042C548.148,157.333,548.148,173.667,548.148,181.833L548.148,190"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_right LE-generate" id="L-gen_sq_right-generate-0" d="M540.648,201.861L529.143,208.551C517.638,215.241,494.628,228.62,483.122,239.477C471.617,250.333,471.617,258.667,471.617,262.833L471.617,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-gen_sq_right LE-include_runtime" id="L-gen_sq_right-include_runtime-0" d="M555.648,201.861L567.154,208.551C578.659,215.241,601.669,228.62,613.174,239.477C624.68,250.333,624.68,258.667,624.68,262.833L624.68,267"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-include_runtime LE-transform_runtime" id="L-include_runtime-transform_runtime-0" d="M624.68,306L624.68,310.167C624.68,314.333,624.68,322.667,624.68,331C624.68,339.333,624.68,347.667,624.68,351.833L624.68,356"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g transform="translate(357.6796875, 141)" class="edgeLabel"><g transform="translate(-71.140625, -12)" class="label"><foreignObject height="24" width="142.28125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel">For every document</span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(357.6796875, 19.5)" id="flowchart-t_before_g-123" class="node default default"><rect height="39" width="208.90625" y="-19.5" x="-104.453125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-96.953125, -12)" style="" class="label"><foreignObject height="24" width="193.90625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_before_generate</span></div></foreignObject></g></g><g transform="translate(357.6796875, 96.5)" id="flowchart-gen_sq_center-124" class="node default default"><rect height="15" width="15" y="-7.5" x="-7.5" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(0, 0)" style="" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel"></span></div></foreignObject></g></g><g transform="translate(124.51171875, 197.5)" id="flowchart-gen_sq_left-126" class="node default default"><rect height="15" width="15" y="-7.5" x="-7.5" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(0, 0)" style="" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel"></span></div></foreignObject></g></g><g transform="translate(41.3046875, 286.5)" id="flowchart-index_file-129" class="node default default"><rect height="39" width="82.609375" y="-19.5" x="-41.3046875" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-33.8046875, -12)" style="" class="label"><foreignObject height="24" width="67.609375"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">index_file</span></div></foreignObject></g></g><g transform="translate(207.71875, 286.5)" id="flowchart-graphql_tag_return-132" class="node default default"><rect height="39" width="150.21875" y="-19.5" x="-75.109375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-67.609375, -12)" style="" class="label"><foreignObject height="24" width="135.21875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">graphql_tag_return</span></div></foreignObject></g></g><g transform="translate(357.6796875, 197.5)" id="flowchart-artifact_data-135" class="node default default"><rect height="39" width="103.0625" y="-19.5" x="-51.53125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-44.03125, -12)" style="" class="label"><foreignObject height="24" width="88.0625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">artifact_data</span></div></foreignObject></g></g><g transform="translate(357.6796875, 286.5)" id="flowchart-artifact_hash-137" class="node default default"><rect height="39" width="49.703125" y="-19.5" x="-24.8515625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-17.3515625, -12)" style="" class="label"><foreignObject height="24" width="34.703125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">hash</span></div></foreignObject></g></g><g transform="translate(357.6796875, 375.5)" id="flowchart-artifact_selection-139" class="node default default"><rect height="39" width="135.0625" y="-19.5" x="-67.53125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-60.03125, -12)" style="" class="label"><foreignObject height="24" width="120.0625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">artifact_selection</span></div></foreignObject></g></g><g transform="translate(357.6796875, 464.5)" id="flowchart-artifact_e-141" class="node default default"><rect height="39" width="98.609375" y="-19.5" x="-49.3046875" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-41.8046875, -12)" style="" class="label"><foreignObject height="24" width="83.609375"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">artifact_end</span></div></foreignObject></g></g><g transform="translate(548.1484375, 197.5)" id="flowchart-gen_sq_right-144" class="node default default"><rect height="15" width="15" y="-7.5" x="-7.5" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(0, 0)" style="" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel"></span></div></foreignObject></g></g><g transform="translate(471.6171875, 286.5)" id="flowchart-generate-147" class="node default default"><rect height="39" width="78.171875" y="-19.5" x="-39.0859375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-31.5859375, -12)" style="" class="label"><foreignObject height="24" width="63.171875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">generate</span></div></foreignObject></g></g><g transform="translate(624.6796875, 286.5)" id="flowchart-include_runtime-150" class="node default default"><rect height="39" width="127.953125" y="-19.5" x="-63.9765625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-56.4765625, -12)" style="" class="label"><foreignObject height="24" width="112.953125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">include_runtime</span></div></foreignObject></g></g><g transform="translate(624.6796875, 375.5)" id="flowchart-transform_runtime-152" class="node default default"><rect height="39" width="144.828125" y="-19.5" x="-72.4140625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-64.9140625, -12)" style="" class="label"><foreignObject height="24" width="129.828125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_runtime</span></div></foreignObject></g></g></g></g></g></svg>
