+++ site/src/routes/_diagrams/rendered/setup.svelte
@@ -1,1 +1,1 @@
-<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 304.6875 55" style="max-width: 304.6875px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674937675514" {...$$props} ><style>#mermaid-1674937675514{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674937675514 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674937675514 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674937675514 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674937675514 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674937675514 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674937675514 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674937675514 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674937675514 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674937675514 .marker.cross{stroke:var(--contrast);}#mermaid-1674937675514 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674937675514 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674937675514 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675514 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675514 .label text,#mermaid-1674937675514 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674937675514 .node rect,#mermaid-1674937675514 .node circle,#mermaid-1674937675514 .node ellipse,#mermaid-1674937675514 .node polygon,#mermaid-1674937675514 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674937675514 .node .label{text-align:center;}#mermaid-1674937675514 .node.clickable{cursor:pointer;}#mermaid-1674937675514 .arrowheadPath{fill:undefined;}#mermaid-1674937675514 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674937675514 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674937675514 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674937675514 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674937675514 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674937675514 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675514 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675514 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674937675514 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674937675514 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-config LE-env" id="L-config-env-0" d="M58.758,19.5L62.924,19.5C67.091,19.5,75.424,19.5,83.758,19.5C92.091,19.5,100.424,19.5,104.591,19.5L108.758,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-env LE-after_load" id="L-env-after_load-0" d="M149.063,19.5L153.229,19.5C157.396,19.5,165.729,19.5,174.063,19.5C182.396,19.5,190.729,19.5,194.896,19.5L199.063,19.5"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(29.37890625, 19.5)" id="flowchart-config-10" class="node default default"><rect height="39" width="58.7578125" y="-19.5" x="-29.37890625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-21.87890625, -12)" style="" class="label"><foreignObject height="24" width="43.7578125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">config</span></div></foreignObject></g></g><g transform="translate(128.91015625, 19.5)" id="flowchart-env-11" class="node default default"><rect height="39" width="40.3046875" y="-19.5" x="-20.15234375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-12.65234375, -12)" style="" class="label"><foreignObject height="24" width="25.3046875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">env</span></div></foreignObject></g></g><g transform="translate(243.875, 19.5)" id="flowchart-after_load-13" class="node default default"><rect height="39" width="89.625" y="-19.5" x="-44.8125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-37.3125, -12)" style="" class="label"><foreignObject height="24" width="74.625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">after_load</span></div></foreignObject></g></g></g></g></g></svg>
+<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 300.671875 55" style="max-width: 300.671875px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674954534827" {...$$props} ><style>#mermaid-1674954534827{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674954534827 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674954534827 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674954534827 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674954534827 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674954534827 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674954534827 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674954534827 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674954534827 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674954534827 .marker.cross{stroke:var(--contrast);}#mermaid-1674954534827 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674954534827 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674954534827 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534827 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534827 .label text,#mermaid-1674954534827 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674954534827 .node rect,#mermaid-1674954534827 .node circle,#mermaid-1674954534827 .node ellipse,#mermaid-1674954534827 .node polygon,#mermaid-1674954534827 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674954534827 .node .label{text-align:center;}#mermaid-1674954534827 .node.clickable{cursor:pointer;}#mermaid-1674954534827 .arrowheadPath{fill:undefined;}#mermaid-1674954534827 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674954534827 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674954534827 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674954534827 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674954534827 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674954534827 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534827 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534827 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674954534827 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674954534827 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-config LE-env" id="L-config-env-0" d="M57.703,19.5L61.87,19.5C66.036,19.5,74.37,19.5,82.703,19.5C91.036,19.5,99.37,19.5,103.536,19.5L107.703,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-env LE-after_load" id="L-env-after_load-0" d="M148.5,19.5L152.667,19.5C156.833,19.5,165.167,19.5,173.5,19.5C181.833,19.5,190.167,19.5,194.333,19.5L198.5,19.5"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(28.8515625, 19.5)" id="flowchart-config-10" class="node default default"><rect height="39" width="57.703125" y="-19.5" x="-28.8515625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-21.3515625, -12)" style="" class="label"><foreignObject height="24" width="42.703125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">config</span></div></foreignObject></g></g><g transform="translate(128.1015625, 19.5)" id="flowchart-env-11" class="node default default"><rect height="39" width="40.796875" y="-19.5" x="-20.3984375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-12.8984375, -12)" style="" class="label"><foreignObject height="24" width="25.796875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">env</span></div></foreignObject></g></g><g transform="translate(241.5859375, 19.5)" id="flowchart-after_load-13" class="node default default"><rect height="39" width="86.171875" y="-19.5" x="-43.0859375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-35.5859375, -12)" style="" class="label"><foreignObject height="24" width="71.171875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">after_load</span></div></foreignObject></g></g></g></g></g></svg>
