+++ site/src/routes/_diagrams/rendered/validate.svelte
@@ -1,1 +1,1 @@
-<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 591.375 55" style="max-width: 591.375px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674937675568" {...$$props} ><style>#mermaid-1674937675568{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674937675568 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674937675568 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674937675568 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674937675568 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674937675568 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674937675568 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674937675568 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674937675568 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674937675568 .marker.cross{stroke:var(--contrast);}#mermaid-1674937675568 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674937675568 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674937675568 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675568 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675568 .label text,#mermaid-1674937675568 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674937675568 .node rect,#mermaid-1674937675568 .node circle,#mermaid-1674937675568 .node ellipse,#mermaid-1674937675568 .node polygon,#mermaid-1674937675568 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674937675568 .node .label{text-align:center;}#mermaid-1674937675568 .node.clickable{cursor:pointer;}#mermaid-1674937675568 .arrowheadPath{fill:undefined;}#mermaid-1674937675568 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674937675568 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674937675568 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674937675568 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674937675568 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674937675568 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674937675568 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674937675568 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674937675568 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674937675568 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-t_before_v LE-validate" id="L-t_before_v-validate-0" d="M206.969,19.5L211.135,19.5C215.302,19.5,223.635,19.5,231.969,19.5C240.302,19.5,248.635,19.5,252.802,19.5L256.969,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-validate LE-t_after_v" id="L-validate-t_after_v-0" d="M329.883,19.5L334.049,19.5C338.216,19.5,346.549,19.5,354.883,19.5C363.216,19.5,371.549,19.5,375.716,19.5L379.883,19.5"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(103.484375, 19.5)" id="flowchart-t_before_v-58" class="node default default"><rect height="39" width="206.96875" y="-19.5" x="-103.484375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-95.984375, -12)" style="" class="label"><foreignObject height="24" width="191.96875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_before_validate</span></div></foreignObject></g></g><g transform="translate(293.42578125, 19.5)" id="flowchart-validate-59" class="node default default"><rect height="39" width="72.9140625" y="-19.5" x="-36.45703125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-28.95703125, -12)" style="" class="label"><foreignObject height="24" width="57.9140625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">validate</span></div></foreignObject></g></g><g transform="translate(477.62890625, 19.5)" id="flowchart-t_after_v-61" class="node default default"><rect height="39" width="195.4921875" y="-19.5" x="-97.74609375" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-90.24609375, -12)" style="" class="label"><foreignObject height="24" width="180.4921875"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_after_validate</span></div></foreignObject></g></g></g></g></g></svg>
+<svg aria-roledescription="flowchart-v2" viewBox="-8 -8 574.578125 55" style="max-width: 574.578125px;" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns="http://www.w3.org/2000/svg" width="100%" id="mermaid-1674954534905" {...$$props} ><style>#mermaid-1674954534905{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;fill:var(--contrast);}#mermaid-1674954534905 .error-icon{fill:hsl(120, 100%, 30.0980392157%);}#mermaid-1674954534905 .error-text{fill:rgb(255, 101.4999999999, 255);stroke:rgb(255, 101.4999999999, 255);}#mermaid-1674954534905 .edge-thickness-normal{stroke-width:2px;}#mermaid-1674954534905 .edge-thickness-thick{stroke-width:3.5px;}#mermaid-1674954534905 .edge-pattern-solid{stroke-dasharray:0;}#mermaid-1674954534905 .edge-pattern-dashed{stroke-dasharray:3;}#mermaid-1674954534905 .edge-pattern-dotted{stroke-dasharray:2;}#mermaid-1674954534905 .marker{fill:var(--contrast);stroke:var(--contrast);}#mermaid-1674954534905 .marker.cross{stroke:var(--contrast);}#mermaid-1674954534905 svg{font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:16px;}#mermaid-1674954534905 .label{font-family:"trebuchet ms",verdana,arial,sans-serif;color:var(--contrast);}#mermaid-1674954534905 .cluster-label text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534905 .cluster-label span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534905 .label text,#mermaid-1674954534905 span{fill:var(--contrast);color:var(--contrast);}#mermaid-1674954534905 .node rect,#mermaid-1674954534905 .node circle,#mermaid-1674954534905 .node ellipse,#mermaid-1674954534905 .node polygon,#mermaid-1674954534905 .node path{fill:var(--discreet2);stroke:var(--contrast);stroke-width:1px;}#mermaid-1674954534905 .node .label{text-align:center;}#mermaid-1674954534905 .node.clickable{cursor:pointer;}#mermaid-1674954534905 .arrowheadPath{fill:undefined;}#mermaid-1674954534905 .edgePath .path{stroke:var(--contrast);stroke-width:2.0px;}#mermaid-1674954534905 .flowchart-link{stroke:var(--contrast);fill:none;}#mermaid-1674954534905 .edgeLabel{background-color:var(--hue);text-align:center;}#mermaid-1674954534905 .edgeLabel rect{opacity:0.5;background-color:var(--hue);fill:var(--hue);}#mermaid-1674954534905 .cluster rect{fill:hsl(120, 100%, 30.0980392157%);stroke:hsl(120, 60%, 20.0980392157%);stroke-width:1px;}#mermaid-1674954534905 .cluster text{fill:rgb(255, 101.4999999999, 255);}#mermaid-1674954534905 .cluster span{color:rgb(255, 101.4999999999, 255);}#mermaid-1674954534905 div.mermaidTooltip{position:absolute;text-align:center;max-width:200px;padding:2px;font-family:"trebuchet ms",verdana,arial,sans-serif;font-size:12px;background:hsl(120, 100%, 30.0980392157%);border:1px solid undefined;border-radius:2px;pointer-events:none;z-index:100;}#mermaid-1674954534905 .flowchartTitleText{text-anchor:middle;font-size:18px;fill:var(--contrast);}#mermaid-1674954534905 :root{--mermaid-font-family:"trebuchet ms",verdana,arial,sans-serif;}</style><g><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="10" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointEnd"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 0 L 10 5 L 0 10 z"></path></marker><marker orient="auto" markerHeight="12" markerWidth="12" markerUnits="userSpaceOnUse" refY="5" refX="0" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-pointStart"><path style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 0 5 L 10 10 L 10 0 z"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="11" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleEnd"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5" refX="-1" viewBox="0 0 10 10" class="marker flowchart" id="flowchart-circleStart"><circle style="stroke-width: 1; stroke-dasharray: 1, 0;" class="arrowMarkerPath" r="5" cy="5" cx="5"></circle></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="12" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossEnd"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><marker orient="auto" markerHeight="11" markerWidth="11" markerUnits="userSpaceOnUse" refY="5.2" refX="-1" viewBox="0 0 11 11" class="marker cross flowchart" id="flowchart-crossStart"><path style="stroke-width: 2; stroke-dasharray: 1, 0;" class="arrowMarkerPath" d="M 1,1 l 9,9 M 10,1 l -9,9"></path></marker><g class="root"><g class="clusters"></g><g class="edgePaths"><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-t_before_v LE-validate" id="L-t_before_v-validate-0" d="M200.891,19.5L205.057,19.5C209.224,19.5,217.557,19.5,225.891,19.5C234.224,19.5,242.557,19.5,246.724,19.5L250.891,19.5"></path><path marker-end="url(#flowchart-pointEnd)" style="fill:none;" class="edge-thickness-normal edge-pattern-solid flowchart-link LS-validate LE-t_after_v" id="L-validate-t_after_v-0" d="M321.047,19.5L325.214,19.5C329.38,19.5,337.714,19.5,346.047,19.5C354.38,19.5,362.714,19.5,366.88,19.5L371.047,19.5"></path></g><g class="edgeLabels"><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g><g class="edgeLabel"><g transform="translate(0, 0)" class="label"><foreignObject height="0" width="0"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="edgeLabel"></span></div></foreignObject></g></g></g><g class="nodes"><g transform="translate(100.4453125, 19.5)" id="flowchart-t_before_v-58" class="node default default"><rect height="39" width="200.890625" y="-19.5" x="-100.4453125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-92.9453125, -12)" style="" class="label"><foreignObject height="24" width="185.890625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_before_validate</span></div></foreignObject></g></g><g transform="translate(285.96875, 19.5)" id="flowchart-validate-59" class="node default default"><rect height="39" width="70.15625" y="-19.5" x="-35.078125" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-27.578125, -12)" style="" class="label"><foreignObject height="24" width="55.15625"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">validate</span></div></foreignObject></g></g><g transform="translate(464.8125, 19.5)" id="flowchart-t_after_v-61" class="node default default"><rect height="39" width="187.53125" y="-19.5" x="-93.765625" ry="0" rx="0" style="" class="basic label-container"></rect><g transform="translate(-86.265625, -12)" style="" class="label"><foreignObject height="24" width="172.53125"><div style="display: inline-block; white-space: nowrap;" xmlns="http://www.w3.org/1999/xhtml"><span class="nodeLabel">transform_after_validate</span></div></foreignObject></g></g></g></g></g></svg>
